04-projects/15_faq_bot/requirements.txt
streamlit
google-genai
chromadb
PyPDF2
python-dotenv